backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
openai==1.51.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
